AsynchronousTask: simplify isAlive (bug 711688)
Simplify all AsynchronousTask subclasses to use the default
isAlive implementation, which returns True if self.returncode
is not None. This fixes cases where the method would erroneously
return False, leading to issues like bug 711688, where the
CompositeTask isAlive implementation returned False for a
BinpkgPrefetcher instance that was still in the process of
starting via its async_start method.

Fixes: d66e9ec0b105 ("AsynchronousTask: add coroutine async_start method")
Bug: https://bugs.gentoo.org/711688

+++ lib/_emerge/SubProcess.py
@@ -24,7 +24,7 @@ class SubProcess(AbstractPollTask):
 		return self.returncode
 
 	def _cancel(self):
-		if self.isAlive():
+		if self.isAlive() and self.pid is not None:
 			try:
 				os.kill(self.pid, signal.SIGTERM)
 			except OSError as e:
@@ -36,10 +36,6 @@ class SubProcess(AbstractPollTask):
 						noiselevel=-1)
 				elif e.errno != errno.ESRCH:
 					raise
-
-	def isAlive(self):
-		return self.pid is not None and \
-			self.returncode is None
 
 	def _async_wait(self):
 		if self.returncode is None:
